New style tests for Pearson
